Updating dynamic inventory files
